+++ crates/holochain/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Add metrics to record incoming notify requests [#5451](https://github.com/holochain/holochain/pull/5451)
+
 ## 0.6.0-rc.2
 
 - Feat: Upgrade Kitsune2 to v0.3.2 \#5449